+++ database/migrations/2016_04_07_161308_make_cache_table.php
@@ -12,11 +12,13 @@ class MakeCacheTable extends Migration
      */
     public function up()
     {
-        Schema::create('cache', function($table) {
-            $table->string('key')->unique();
-            $table->text('value');
-            $table->integer('expiration');
-        });
+        if (!Schema::hasTable('cache')) {
+            Schema::create('cache', function ($table) {
+                $table->string('key')->unique();
+                $table->text('value');
+                $table->integer('expiration');
+            });
+        }
     }
 
     /**
@@ -26,6 +28,8 @@ class MakeCacheTable extends Migration
      */
     public function down()
     {
-        Schema::drop('cache');
+        if (Schema::hasTable('cache')) {
+            Schema::drop('cache');
+        }
     }
 }
