+++ database/migrations/2016_04_16_013749_update_updates_table_with_geo_location.php
@@ -12,10 +12,12 @@ class UpdateUpdatesTableWithGeoLocation extends Migration
      */
     public function up()
     {
-        Schema::table('update_recipients', function(Blueprint $table) {
-           $table->string('geo_lat');
-           $table->string('geo_long');
-        });
+        if (Schema::hasTable('update_recipients') and !Schema::hasColumns('update_recipients', ['geo_lat','geo_long'])) {
+            Schema::table('update_recipients', function (Blueprint $table) {
+                $table->string('geo_lat');
+                $table->string('geo_long');
+            });
+        }
     }
 
     /**
@@ -25,9 +27,11 @@ class UpdateUpdatesTableWithGeoLocation extends Migration
      */
     public function down()
     {
-        Schema::table('update_recipients', function(Blueprint $table) {
-            $table->dropColumn('geo_lat');
-            $table->dropColumn('geo_long');
-        });
+        if (Schema::hasTable('update_recipients') and Schema::hasColumns('update_recipients', ['geo_lat','geo_long'])) {
+            Schema::table('update_recipients', function (Blueprint $table) {
+                $table->dropColumn('geo_lat');
+                $table->dropColumn('geo_long');
+            });
+        }
     }
 }
