+++ database/migrations/2016_10_31_200100_add_confirmation_column_to_form_mail_table.php
@@ -12,9 +12,11 @@ class AddConfirmationColumnToFormMailTable extends Migration
      */
     public function up()
     {
-        Schema::table('form_mail', function (Blueprint $table) {
-            $table->boolean('confirmation');
-        });
+        if (!Schema::hasColumn('form_mail', 'confirmation')) {
+            Schema::table('form_mail', function (Blueprint $table) {
+                $table->boolean('confirmation');
+            });
+        }
     }
 
     /**
@@ -24,8 +26,10 @@ class AddConfirmationColumnToFormMailTable extends Migration
      */
     public function down()
     {
-        Schema::table('form_mail', function (Blueprint $table) {
-            $table->dropColumn('confirmation');
-        });
+        if (Schema::hasColumn('form_mail', 'confirmation')) {
+            Schema::table('form_mail', function (Blueprint $table) {
+                $table->dropColumn('confirmation');
+            });
+        }
     }
 }
