+++ database/migrations/2016_10_31_201854_add_queue_column_to_form_mail_table.php
@@ -12,9 +12,11 @@ class AddQueueColumnToFormMailTable extends Migration
      */
     public function up()
     {
-        Schema::table('form_mail', function (Blueprint $table) {
-            $table->boolean('queue');
-        });
+        if (!Schema::hasColumn('form_mail','queue')) {
+            Schema::table('form_mail', function (Blueprint $table) {
+                $table->boolean('queue');
+            });
+        }
     }
 
     /**
@@ -24,8 +26,10 @@ class AddQueueColumnToFormMailTable extends Migration
      */
     public function down()
     {
-        Schema::table('form_mail', function (Blueprint $table) {
-            $table->dropColumn('queue');
-        });
+        if (Schema::hasColumn('form_mail','queue')) {
+            Schema::table('form_mail', function (Blueprint $table) {
+                $table->dropColumn('queue');
+            });
+        }
     }
 }
